+++ asta4d-core/src/main/java/com/astamuse/asta4d/util/IdGenerator.java
@@ -19,12 +19,16 @@ package com.astamuse.asta4d.util;
 
 import java.nio.ByteBuffer;
 import java.security.SecureRandom;
-
-import org.apache.commons.lang3.StringUtils;
+import java.util.Base64;
+import java.util.Base64.Encoder;
 
 public class IdGenerator {
 
-    private final static int randomSeed;
+    private static Encoder b64Encoder = Base64.getUrlEncoder();
+
+    // this seed must by 5 bytes due to avoid base64 padding later
+    private final static byte[] randomSeed;
+
     static {
         // use the last 32 bit of current time as the seed
         ByteBuffer bb = ByteBuffer.allocate(64);
@@ -32,23 +36,35 @@ public class IdGenerator {
         byte[] bytes = bb.array();
         byte[] seed = new byte[4];
         System.arraycopy(bytes, 4, seed, 0, 4);
+
         SecureRandom sr = new SecureRandom(seed);
-        randomSeed = Math.abs(sr.nextInt());
+        int s = Math.abs(sr.nextInt());
+
+        ByteBuffer seedBuffer = ByteBuffer.allocate(5);
+        seedBuffer.putInt(s);
+        seedBuffer.put(seed[3]);
+
+        randomSeed = seedBuffer.array();
     }
 
     private final static class IdHolder {
+
+        // must be times of 3 to avoid padding, 8 + 8 + 5 = 21
+        private ByteBuffer buffer = ByteBuffer.allocate(21);
+
         private long threadId;
         private long lastTime = Long.MIN_VALUE;
 
         public IdHolder(long threadId) {
             this.threadId = threadId;
+            this.buffer.mark();
         }
 
         public long getThreadId() {
             return this.threadId;
         }
 
-        public long newTime() {
+        long newTime() {
             // since the current milliseconds is less than 40 bit, we think this
             // operation is safe
             long cur = System.currentTimeMillis() << 7;
@@ -59,6 +75,16 @@ public class IdGenerator {
                 cur = lastTime;
             }
             return cur;
+        }
+
+        public byte[] newId() {
+            buffer.reset();
+            buffer.putLong(newTime());// 8
+            buffer.putLong(threadId);// 8
+            buffer.put(randomSeed);// 5
+            byte[] bs = new byte[21];
+            System.arraycopy(buffer.array(), 0, bs, 0, 21);
+            return bs;
         }
     }
 
@@ -72,15 +98,22 @@ public class IdGenerator {
     };
 
     /**
-     * a unique id with thread id embedded and a thread unique number
+     * a unique id with thread id embedded and a process unique(random) number, as string.
      * 
      * @return
      */
-    // TODO we want a fast uuid solution
     public final static String createId() {
         IdHolder idHolder = idHolderCache.get();
-        long time = idHolder.newTime();
-        Object[] vals = { time, idHolder.getThreadId(), randomSeed };
-        return StringUtils.join(vals, "-");
+        return b64Encoder.encodeToString(idHolder.newId());
+    }
+
+    /**
+     * a unique id with thread id embedded and a process unique(random) number, as byte array
+     * 
+     * @return
+     */
+    public final static byte[] createIdBytes() {
+        IdHolder idHolder = idHolderCache.get();
+        return idHolder.newId();
     }
 }