+++ asta4d-core/src/test/java/com/astamuse/asta4d/test/render/ParallelTest.java
@@ -22,36 +22,21 @@ import java.util.List;
 
 import org.testng.annotations.Test;
 
+import com.astamuse.asta4d.Configuration;
 import com.astamuse.asta4d.render.Renderer;
 import com.astamuse.asta4d.test.render.infra.BaseTest;
 import com.astamuse.asta4d.test.render.infra.SimpleCase;
 import com.astamuse.asta4d.test.render.infra.TimeCalculator;
 import com.astamuse.asta4d.util.ElementUtil;
-import com.astamuse.asta4d.util.collection.ParallelRowConvertor;
 import com.astamuse.asta4d.util.collection.RowConvertorBuilder;
 
 @Test(singleThreaded = true)
 public class ParallelTest extends BaseTest {
 
     public static class TestRender {
-        public Renderer listTextRendering() {
-            List<String> list = Arrays.asList("a", "b", "c", "d");
-            Renderer renderer = Renderer.create("div#test", list, new ParallelRowConvertor<String, String>() {
-                @Override
-                public String convert(int rowIndex, String obj) {
-                    try {
-                        Thread.sleep(1000);
-                    } catch (InterruptedException e) {
-                        throw new RuntimeException(e);
-                    }
-                    return obj + "-sleep";
-                }
-
-            });
-            return renderer;
-        }
 
         public Renderer listTextRenderingLambda() {
+            Configuration.getConfiguration().setNumberLimitOfParallelListConverting(50);
             List<String> list = Arrays.asList("a", "b", "c", "d");
             Renderer renderer = Renderer.create("div#test", list, RowConvertorBuilder.parallel(obj -> {
                 try {
@@ -93,20 +78,6 @@ public class ParallelTest extends BaseTest {
             }
             return renderer;
         }
-    }
-
-    @Test
-    public void paralleListConversion() {
-        TimeCalculator.shouldRunInTime(new Runnable() {
-            @Override
-            public void run() {
-                try {
-                    new SimpleCase("ParallelTest_parallelListConversion.html");
-                } catch (Throwable t) {
-                    throw new RuntimeException(t);
-                }
-            }
-        }, 2000);
     }
 
     @Test