+++ asta4d-spring/src/main/java/com/astamuse/asta4d/misc/spring/mvc/SpringWebPageViewResolver.java
@@ -21,20 +21,53 @@ import java.util.Locale;
 
 import org.springframework.web.servlet.View;
 import org.springframework.web.servlet.ViewResolver;
+import org.springframework.web.servlet.view.RedirectView;
 
-import com.astamuse.asta4d.Page;
-import com.astamuse.asta4d.template.TemplateException;
-import com.astamuse.asta4d.web.dispatch.response.provider.Asta4DPageProvider;
+import com.astamuse.asta4d.Configuration;
+import com.astamuse.asta4d.template.Template;
+import com.astamuse.asta4d.template.TemplateNotFoundException;
+import com.astamuse.asta4d.template.TemplateResolver;
 
 public class SpringWebPageViewResolver implements ViewResolver {
+
+    private boolean exceptionOnTemplateNotFound = true;
+    private String prefix = "";
+    private String suffix = "";
 
     @Override
     public View resolveViewName(String viewName, Locale locale) throws Exception {
         try {
-            return new SpringWebPageView(new Asta4DPageProvider(Page.buildFromPath(viewName)));
-        } catch (TemplateException e) {
-            return null;
+            // TODO we should follow the spring mvc default implementation of UrlBasedViewResolver
+            String redirectPrefix = "redirect:";
+            if (viewName.startsWith(redirectPrefix)) {
+                String redirect = viewName.substring(redirectPrefix.length());
+                return new RedirectView(redirect);
+            }
+
+            String path = prefix + viewName + suffix;
+            Configuration conf = Configuration.getConfiguration();
+            TemplateResolver templateResolver = conf.getTemplateResolver();
+            Template template = templateResolver.findTemplate(path);
+            return new SpringWebPageView(template);
+        } catch (TemplateNotFoundException e) {
+            if (exceptionOnTemplateNotFound) {
+                throw e;
+            } else {
+                return null;
+            }
         }
+    }
+
+    public void setExceptionOnTemplateNotFound(boolean exceptionOnTemplateNotFound) {
+        this.exceptionOnTemplateNotFound = exceptionOnTemplateNotFound;
+    }
+
+    public void setPrefix(String prefix) {
+        this.prefix = prefix;
+    }
+
+    public void setSuffix(String suffix) {
+        this.suffix = suffix;
     }
 
 }