+++ asta4d-web/src/main/java/com/astamuse/asta4d/web/util/timeout/DefaultSessionAwareExpirableDataManager.java
@@ -17,8 +17,10 @@
 
 package com.astamuse.asta4d.web.util.timeout;
 
+import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
 import java.util.Map.Entry;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.Executors;
@@ -34,9 +36,9 @@ import com.astamuse.asta4d.web.WebApplicationContext;
 
 public class DefaultSessionAwareExpirableDataManager implements ExpirableDataManager {
 
-    private static final String CheckSessionIdKey = DefaultSessionAwareExpirableDataManager.class + "#CheckSessionIdKey";
+    private static final String SessionCheckIdKey = DefaultSessionAwareExpirableDataManager.class + "#SessionCheckIdKey";
 
-    private ConcurrentHashMap<String, DataHolder> dataMap = null;
+    private Map<String, DataHolder> dataMap = null;
 
     private AtomicInteger dataCounter = null;
 
@@ -57,7 +59,8 @@ public class DefaultSessionAwareExpirableDataManager implements ExpirableDataMan
     private String checkThreadName = this.getClass().getSimpleName() + "-check-thread";
 
     public DefaultSessionAwareExpirableDataManager() {
-
+        dataMap = createThreadSafeDataMap();
+        dataCounter = new AtomicInteger();
     }
 
     public void setExpirationCheckPeriodInMilliseconds(long expirationCheckPeriodInMilliseconds) {
@@ -80,11 +83,28 @@ public class DefaultSessionAwareExpirableDataManager implements ExpirableDataMan
         this.maxSpinTimeInMilliseconds = maxSpinTimeInMilliseconds;
     }
 
+    protected Map<String, DataHolder> createThreadSafeDataMap() {
+        return new ConcurrentHashMap<>();
+    }
+
+    protected void decreaseCount() {
+        dataCounter.decrementAndGet();
+    }
+
+    protected void addCount(int delta) {
+        dataCounter.addAndGet(delta);
+    }
+
+    protected void increaseCount() {
+        dataCounter.incrementAndGet();
+    }
+
+    protected int getCurrentCount() {
+        return dataCounter.get();
+    }
+
     @Override
     public void start() {
-        // init fields
-        dataMap = new ConcurrentHashMap<>();
-        dataCounter = new AtomicInteger();
         service = Executors.newSingleThreadScheduledExecutor(new ThreadFactory() {
             @Override
             public Thread newThread(Runnable r) {
@@ -109,7 +129,7 @@ public class DefaultSessionAwareExpirableDataManager implements ExpirableDataMan
                     }
                 }
                 if (removedCounter > 0) {
-                    dataCounter.addAndGet(-removedCounter);
+                    addCount(-removedCounter);
                 }
             }
         }, expirationCheckPeriodInMilliseconds, expirationCheckPeriodInMilliseconds, TimeUnit.MILLISECONDS);
@@ -129,7 +149,7 @@ public class DefaultSessionAwareExpirableDataManager implements ExpirableDataMan
         if (remove) {
             holder = dataMap.remove(dataId);
             if (holder != null) {
-                dataCounter.decrementAndGet();
+                decreaseCount();
                 if (holder.isExpired(System.currentTimeMillis())) {
                     holder = null;
                 }
@@ -140,7 +160,7 @@ public class DefaultSessionAwareExpirableDataManager implements ExpirableDataMan
                 if (holder.isExpired(System.currentTimeMillis())) {
                     holder = dataMap.remove(dataId);
                     if (holder != null) {
-                        dataCounter.decrementAndGet();
+                        decreaseCount();
                         holder = null;
                     }
                 }
@@ -149,7 +169,7 @@ public class DefaultSessionAwareExpirableDataManager implements ExpirableDataMan
         if (holder == null) {
             return null;
         } else {
-            if (StringUtils.equals(retrieveSessionId(false), holder.sessionId)) {
+            if (StringUtils.equals(retrieveSessionCheckId(false), holder.sessionId)) {
                 return (T) holder.getData();
             } else {
                 return null;
@@ -158,13 +178,21 @@ public class DefaultSessionAwareExpirableDataManager implements ExpirableDataMan
     }
 
     public void put(String dataId, Object data, long expireMilliSeconds) {
-        if (dataCounter.get() >= maxDataSize) {
+        checkSize();
+        Object existing = dataMap.put(dataId, new DataHolder(data, expireMilliSeconds, retrieveSessionCheckId(true)));
+        if (existing == null) {
+            increaseCount();
+        }
+    }
+
+    protected void checkSize() {
+        if (getCurrentCount() >= maxDataSize) {
             try {
                 long spinTimeTotal = 0;
-                while (dataCounter.get() >= maxDataSize) {
+                while (getCurrentCount() >= maxDataSize) {
                     if (spinTimeTotal >= maxSpinTimeInMilliseconds) {
-                        String msg = "There are too many data in %s and we could not get empty space after waiting for %d milliseconds."
-                                + " The configured max size is %d and perhaps you should increase the value.";
+                        String msg = "There are too many data in %s and we could not get empty space after waiting for %d milliseconds." +
+                                " The configured max size is %d and perhaps you should increase the value.";
                         msg = String.format(msg, this.getClass().getName(), spinTimeTotal, maxDataSize);
                         throw new TooManyDataException(msg);
                     }
@@ -175,23 +203,26 @@ public class DefaultSessionAwareExpirableDataManager implements ExpirableDataMan
                 throw new RuntimeException(e);
             }
         }
-        Object existing = dataMap.put(dataId, new DataHolder(data, expireMilliSeconds, retrieveSessionId(true)));
-        if (existing == null) {
-            dataCounter.incrementAndGet();
-        }
     }
 
-    protected String retrieveSessionId(boolean create) {
-        String sessionId = null;
+    /**
+     * NOTE: Because there is no way to retrieve the session id via WebApplicationContext, thus we use a independent check id which is
+     * different from the http request session id to check whether current request client is the same client from previous request.
+     * 
+     * @param create
+     * @return
+     */
+    protected String retrieveSessionCheckId(boolean create) {
+        String sessionCheckId = null;
         if (sessionAware) {
             WebApplicationContext context = WebApplicationContext.getCurrentThreadWebApplicationContext();
-            sessionId = context.getData(WebApplicationContext.SCOPE_SESSION, CheckSessionIdKey);
-            if (sessionId == null && create) {
-                sessionId = IdGenerator.createId();
-                context.setData(WebApplicationContext.SCOPE_SESSION, CheckSessionIdKey, sessionId);
+            sessionCheckId = context.getData(WebApplicationContext.SCOPE_SESSION, SessionCheckIdKey);
+            if (sessionCheckId == null && create) {
+                sessionCheckId = IdGenerator.createId();
+                context.setData(WebApplicationContext.SCOPE_SESSION, SessionCheckIdKey, sessionCheckId);
             }
         }
-        return sessionId;
+        return sessionCheckId;
     }
 
     @Override
@@ -200,7 +231,11 @@ public class DefaultSessionAwareExpirableDataManager implements ExpirableDataMan
         super.finalize();
     }
 
-    private static class DataHolder {
+    protected static class DataHolder implements Serializable {
+        /**
+         * 
+         */
+        private static final long serialVersionUID = 1L;
         private final Object data;
         private final long creationTime;
         private final long expireMilliSeconds;